# Speed up reading uniqState