--- a/tests/performance/uniq_stored.xml
+++ b/tests/performance/uniq_stored.xml
@@ -1,39 +1,58 @@
 <test>
     <create_query>
-        create table landing (
-            t DateTime,
+        create table matview_1
+        (
             a String,
-            b String
-        ) Engine=MergeTree partition by tuple() order by a settings index_granularity = 1024;
+            b_count AggregateFunction(uniq, UInt64)
+        ) Engine=MergeTree partition by tuple()
+        ORDER by tuple()
+        SETTINGS index_granularity = 1024;
     </create_query>
+
     <create_query>
-        create table matview (
+        create table matview_10000
+        (
             a String,
             b_count AggregateFunction(uniq, String)
-        ) Engine=AggregatingMergeTree partition by tuple() order by a settings index_granularity = 1024;
-    </create_query>
-    <create_query>
-        CREATE MATERIALIZED VIEW mv TO matview AS (
-            SELECT a, uniqState(b) b_count
-            FROM landing
-            GROUP BY a
-        );
+        ) Engine=MergeTree partition by tuple()
+        ORDER by tuple()
+        SETTINGS index_granularity = 1024;
     </create_query>
 
-    <drop_query>DROP TABLE IF EXISTS mv</drop_query>
-    <drop_query>DROP TABLE IF EXISTS matview</drop_query>
-    <drop_query>DROP TABLE IF EXISTS landing</drop_query>
+
+    <drop_query>DROP TABLE IF EXISTS matview_1</drop_query>
+    <drop_query>DROP TABLE IF EXISTS matview_10000</drop_query>
 
     <fill_query>
-        insert into landing	select today()-1, toString(rand()%1000),	concat('usr',toString(number%500)) from numbers(10000000);
+        INSERT INTO matview_10000
+        SELECT a, uniqState(b) b_count
+        FROM
+        (
+            SELECT toString(rand() % 1000) a, toString(number % 10000) b
+            FROM numbers_mt(20000000)
+        )
+        GROUP BY a
+        SETTINGS max_insert_threads=8;
     </fill_query>
-    <fill_query>
-        insert into landing select today()-2, toString(rand()%1000), concat('usr',toString(number%400)) from numbers(10000000);
-    </fill_query>
-    <fill_query>
-        insert into landing select today()-3, toString(rand()%1000), concat('usr',toString(number%200)) from numbers(10000000);
-    </fill_query>
+    <fill_query>OPTIMIZE TABLE matview_10000 FINAL</fill_query>
 
-    <query>select a, uniqMerge(b_count) as b_count from matview prewhere a='555' group by a FORMAT Null SETTINGS max_threads=1;</query>
-    <query>select uniqMerge(b_count) as b_count from matview FORMAT Null SETTINGS max_threads=1;</query>
+    <fill_query>
+        INSERT INTO matview_1
+        SELECT '1', uniqState(number) b_count
+        FROM
+        (
+            SELECT *
+            FROM numbers_mt(2000000)
+        )
+        GROUP BY number
+        SETTINGS max_insert_threads=8;
+    </fill_query>
+    <fill_query>OPTIMIZE TABLE matview_1 FINAL</fill_query>
+
+    <!-- Test with ~10000 elements per state -->
+    <query>select a, uniqMerge(b_count) as b_count from matview_10000 prewhere a='55' group by a FORMAT Null SETTINGS max_threads=1;</query>
+    <query>select uniqMerge(b_count) as b_count from matview_10000 FORMAT Null SETTINGS max_threads=1;</query>
+
+    <!-- Test with ~1 elements per state -->
+    <query>select uniqMerge(b_count) as b_count FROM matview_1 FORMAT Null SETTINGS max_threads=1;</query>
 </test>