### Changelog category (leave one):
- Performance Improvement

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Speed up reading UniquesHashSet (uniqState from disk for example)

> Information about CI checks: https://clickhouse.com/docs/en/development/continuous-integration/

In some cases, the uniqMerge process for a query was completely dominated by the many calls to read data from the buffer into a uniqState state. Let's read all the integers in a block to avoid the overhead.


Example:

* Setup:
```sql
select version();

drop database if exists tt;

create database tt;

create table tt.landing (
	t DateTime,
	a String,
	b String
) Engine=MergeTree partition by tuple() order by a settings index_granularity = 1024;

create table tt.matview (
	a String,
	b_count AggregateFunction(uniq, String)
) Engine=AggregatingMergeTree partition by tuple() order by a settings index_granularity = 1024;

CREATE MATERIALIZED VIEW tt.mv TO tt.matview AS (
	SELECT a, uniqState(b) b_count
	FROM tt.landing
	GROUP BY a
);

insert into tt.landing select today()-1, toString(rand()%1000), concat('usr',toString(number%500)) from numbers(10000000);
insert into tt.landing select today()-2, toString(rand()%1000), concat('usr',toString(number%400)) from numbers(10000000);
insert into tt.landing select today()-3, toString(rand()%1000), concat('usr',toString(number%200)) from numbers(10000000);
```

* `clickhouse-benchmark --query "select a, uniqMerge(b_count) as b_count from tt.matview prewhere a='555' group by a FORMAT Null SETTINGS max_threads=1;" -i 2000`
  - Master: localhost:9000, queries 2000, QPS: 151.412, RPS: 151411.945, MiB/s: 21.207, result RPS: 0.000, result MiB/s: 0.000.
  - Changes: localhost:9000, queries 2000, QPS: 282.340, RPS: 282339.980, MiB/s: 39.545, result RPS: 0.000, result MiB/s: 0.000.

* `clickhouse-benchmark --query "select a, uniqMerge(b_count) as b_count from tt.matview group by a FORMAT Null SETTINGS max_threads=1;" -i 1000`
  - Master: localhost:9000, queries 1000, QPS: 81.189, RPS: 81188.559, MiB/s: 11.371, result RPS: 0.000, result MiB/s: 0.000.
  - Changes: localhost:9000, queries 1000, QPS: 105.919, RPS: 105919.390, MiB/s: 14.835, result RPS: 0.000, result MiB/s: 0.000.