@Algunenano 2022-09-07
Thinking out loud late at night: I wonder if this needs to handle endianness changes (where the generator of the data and the reader have different endianness). I haven't seen anything like that anywhere so I'd assume it's not supported in general, but I haven't looked much yet to confirm that's the case