@Algunenano 2022-09-09
Replying to myself: No other part of the code seems to consider endianness changes anywhere and there are similar approaches in statuses like `RoaringBitmapWithSmallSet` or `QuantileTimingTiny`. Also, there isn't any conversion in ReadHelpers.h (`readIntBinary` et. al) which would be the place where I'd add it.

So unless somebody says the opposite I'd say it's safe to assume you can't use data (via storage or network) generated by CH servers with a different endianness as the current server.