@Algunenano 2022-09-09
Looking at test errors:
- [Stateless tests (debug, s3 storage) [1/3]](https://s3.amazonaws.com/clickhouse-test-reports/41089/eae48c273b5bb46a350549a00a4fb10b08697ecd/stateless_tests__debug__s3_storage__[1/3].html) -> 00632_get_sample_block_cache. A timeout, likely related to s3 connection failing?
- [Stateless tests (release, s3 storage)](https://s3.amazonaws.com/clickhouse-test-reports/41089/eae48c273b5bb46a350549a00a4fb10b08697ecd/stateless_tests__release__s3_storage_.html) -> 01396_inactive_replica_cleanup_nodes_zookeeper. Seems like a s3 error: `AWSErrorMarshaller: Encountered AWSError 'NoSuchKey': The specified key does not exist.`
- [Stateless tests (tsan) [3/3]](https://s3.amazonaws.com/clickhouse-test-reports/41089/eae48c273b5bb46a350549a00a4fb10b08697ecd/stateless_tests__tsan__[3/3].html) -> `Server does not respond to health check`. The server took too long to reply so the test handler thought it was dead (but it wasn't) and stopped testing. No idea why the query (`4c155115-283b-48fb-9400-90cee8a267ac` `SELECT 'Running test stateless/01698_fix_toMinute.sql from pid=940'` took almost 2 minutes to finish).